Finished redis based session authentication

+++ graph/routes/Article.go
@@ -20,14 +20,13 @@ func CreateArticle(input *model.CreateArticleInfo) (*model.Article, error) {
 		Creates a temporary array for the article model, loop through the contents of the input for all the tag data
 	*/
 	message, _ := JWTValidityCheck(input.Jwt)
-	if message == "Unauthorized!" {
-		panic("Unauthorized!")
-	}
-
 	redisClient := RedisClientInstation()
 	redisData := RedisUserInfo(input.Jwt, redisClient)
+	if message == "Unauthorized!" || redisData["role"] == "Reader" {
+		panic("Unauthorized!")
+	}
 	username := redisData["username"]
-	password := redisData["password"]
+	zincUsername, zincPassword := ZincLogin(input.ProjectUUID)
 	var tags []model.Tag
 	var tagsString []string
 	for tagData := 0; tagData < len(input.Tags); tagData++ {
@@ -39,7 +38,7 @@ func CreateArticle(input *model.CreateArticleInfo) (*model.Article, error) {
 	}
 	imageURL := UploadFileToS3(input)
 	client := ConnectToMongo()
-	collection := client.Database(fmt.Sprintf("%s-%s", username, input.ProjectUUID)).Collection("articles")
+	collection := client.Database(fmt.Sprintf("%s", input.ProjectUUID)).Collection("articles")
 	author := model.Author{Name: username}
 	article := model.Article{Title: *input.Title, Author: &author, ContentData: *input.ContentData, DateWritten: *input.DateWritten, URL: *input.URL, Description: *input.Description, UUID: *input.UUID, Tags: tags, TitleCard: imageURL}
 	res, err := collection.InsertOne(context.TODO(), article)
@@ -66,14 +65,13 @@ func CreateArticle(input *model.CreateArticleInfo) (*model.Article, error) {
 	log.WithFields(log.Fields{
 		"article state": "created mongodb instance",
 	}).Info("Article has been created, inserting into zinc!")
-
-	CreateDocument(fmt.Sprintf("%s-%s-articles", username, input.ProjectUUID), zincData, *input.UUID, username, password)
+	CreateDocument(fmt.Sprintf("%s-articles", input.ProjectUUID), zincData, input.ProjectUUID, zincUsername, zincPassword)
 	log.WithFields(log.Fields{
 		"article state": "finished insertion",
 	}).Info(fmt.Sprintf("Inserted a single document: %s", res.InsertedID))
 	return &article, err
 }
-func DeleteArticle(bucket *model.DeleteBucketInfo) (*model.Article, error) {
+func DeleteArticle(bucket *model.DeleteBucketInfo) (string, error) {
 	message, _ := JWTValidityCheck(bucket.Jwt)
 	if message == "Unauthorized!" {
 		panic("Unauthorized!")
@@ -81,39 +79,43 @@ func DeleteArticle(bucket *model.DeleteBucketInfo) (*model.Article, error) {
 	redisClient := RedisClientInstation()
 	redisData := RedisUserInfo(bucket.Jwt, redisClient)
 	username := redisData["username"]
-	password := redisData["password"]
+	zincusername, password := ZincLogin(bucket.ProjectUUID)
+	role := redisData["role"]
 	client := ConnectToMongo()
-	collection := client.Database(fmt.Sprintf("%s-%s", username, bucket.ProjectUUID)).Collection("articles")
-	article := model.Article{UUID: *bucket.UUID}
-	session := CreateAWSSession()
-	s3sc := s3.New(session)
-	bucketName := fmt.Sprintf("%s-%s-images", username, bucket.ProjectUUID)
-	iter := s3manager.NewDeleteListIterator(s3sc, &s3.ListObjectsInput{
-		Bucket: aws.String(bucketName),
-		Prefix: &bucket.Articlename,
-	})
+	if role == "Admin" || username == bucket.Username {
+		collection := client.Database(fmt.Sprintf("%s", bucket.ProjectUUID)).Collection("articles")
+		session := CreateAWSSession()
+		s3sc := s3.New(session)
+		bucketName := fmt.Sprintf("%s-%s-images", username, bucket.ProjectUUID)
+		iter := s3manager.NewDeleteListIterator(s3sc, &s3.ListObjectsInput{
+			Bucket: aws.String(bucketName),
+			Prefix: &bucket.Articlename,
+		})
 
-	DeleteArticleFolder(s3sc, iter, bucketName)
-	deleteResult, deleteError := collection.DeleteOne(context.TODO(), bson.M{"uuid": *bucket.UUID})
-	if deleteResult.DeletedCount == 0 {
-		log.Fatal("Error on deleting data ", deleteError)
+		DeleteArticleFolder(s3sc, iter, bucketName)
+		deleteResult, deleteError := collection.DeleteOne(context.TODO(), bson.M{"uuid": *bucket.UUID})
+		if deleteResult.DeletedCount == 0 {
+			log.Fatal("Error on deleting data ", deleteError)
+		}
+		zincData := fmt.Sprintf(`{
+			"UUID":        "%s"
+		}`, *bucket.UUID)
+
+		DeleteDocument(fmt.Sprintf("%s-%s-articles", username, bucket.ProjectUUID), zincData, bucket.ProjectUUID, zincusername, password)
+		return "successful", deleteError
 	}
-	zincData := fmt.Sprintf(`{
-		"UUID":        "%s"
-	}`, *bucket.UUID)
-	DeleteDocument(fmt.Sprintf("%s-%s-articles-%s", username, bucket.ProjectUUID, *bucket.UUID), zincData, *bucket.UUID, username, password)
-	return &article, deleteError
+	var err error
+	return "", err
 }
 func FindArticle(input *model.FindArticlePrivateType) (*model.Article, error) {
 	message, _ := JWTValidityCheck(input.Jwt)
 	if message == "Unauthorized!" {
 		panic("Unauthorized!")
 	}
-	redisClient := RedisClientInstation()
-	redisData := RedisUserInfo(input.Jwt, redisClient)
-	username := redisData["username"]
+	// redisClient := RedisClientInstation()
+	// redisData := RedisUserInfo(input.Jwt, redisClient)
 	client := ConnectToMongo()
-	collection := client.Database(fmt.Sprintf("%s-%s", username, input.ProjectUUID)).Collection("articles")
+	collection := client.Database(fmt.Sprintf("%s", input.ProjectUUID)).Collection("articles")
 	var article model.Article
 
 	//Passing the bson.D{{}} as the filter matches documents in the collection
@@ -125,13 +127,12 @@ func FindArticle(input *model.FindArticlePrivateType) (*model.Article, error) {
 }
 func UpdateArticle(input *model.UpdatedArticleInfo) (*model.Article, error) {
 	message, _ := JWTValidityCheck(input.Jwt)
-	if message == "Unauthorized!" {
-		panic("Unauthorized!")
-	}
 	redisClient := RedisClientInstation()
 	redisData := RedisUserInfo(input.Jwt, redisClient)
-	username := redisData["username"]
-	password := redisData["password"]
+	if message == "Unauthorized!" || redisData["role"] == "Reader" {
+		panic("Unauthorized!")
+	}
+	username, password := ZincLogin(input.ProjectUUID)
 	var tags []model.Tag
 	var tagsString []string
 	for tagData := 0; tagData < len(input.Tags); tagData++ {
@@ -143,7 +144,7 @@ func UpdateArticle(input *model.UpdatedArticleInfo) (*model.Article, error) {
 	}
 	imageURL := UploadUpdatedFileToS3(input)
 	client := ConnectToMongo()
-	collection := client.Database(fmt.Sprintf("%s-%s", username, input.ProjectUUID)).Collection("articles")
+	collection := client.Database(fmt.Sprintf("%s", input.ProjectUUID)).Collection("articles")
 
 	filter := bson.M{"uuid": input.UUID}
 	update := bson.D{primitive.E{Key: "$set", Value: bson.D{
@@ -170,6 +171,6 @@ func UpdateArticle(input *model.UpdatedArticleInfo) (*model.Article, error) {
 		"TitleCard":   "%s",
 		"Tags":        "%s"
 	}`, *input.Title, *input.Author, *input.ContentData, *input.DateWritten, *input.URL, *input.Description, *input.UUID, imageURL, strings.Join(tagsString, ","))
-	UpdateDocument("articles", zincData, *input.UUID, username, password)
+	UpdateDocument(fmt.Sprintf("%s-articles", input.ProjectUUID), zincData, *input.UUID, username, password)
 	return &article, ArticleUpdateerr
 }
